third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDiffuseLightingElement-lighting-color-css-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'lighting-color' css property of the SVGFEDiffuseLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.defaultView.getComputedStyle(gradientElement).getPropertyValue('lighting-color') is "rgb(255, 0, 0)"
PASS document.defaultView.getComputedStyle(gradientElement).getPropertyValue('lighting-color') is "rgb(255, 255, 0)"
PASS successfullyParsed is true

TEST COMPLETE
